Inline ElementsClassInspector and move off of immutable nodes API
