--- interop/kotlinx-metadata/classinspectors/elements/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=kotlinpoet-classinspector-elements
-POM_NAME=KotlinPoet (Elements ClassInspector implementation)
-POM_DESCRIPTION=Elements ClassInspector implementation.
-POM_PACKAGING=jar
